refactor: improve swagger PaginatedEntity generic generation
