+++ apps/cms-admin/scripts/swagger-api-template/route-types.ejs
@@ -9,7 +9,7 @@ import { <%~ dataContracts.join(", ") %> } from "./<%~ config.fileNames.dataCont
 <% } %>
 
 <%
-/* TODO: outOfModule, combined should be attributes of route, which will allow to avoid duplication of code */
+/* T0D0: outOfModule, combined should be attributes of route, which will allow to avoid duplication of code */
 %>
 
 <% if (routes.outOfModule) { %>
