--- apps/cms-admin/src/client/type.ts
@@ -1,3 +1,0 @@
-import type { PaginatedMetadata } from 'src/client/cms/cms-api'
-
-export interface PaginatedEntity<T = never> { items: T[], metadata: PaginatedMetadata }
