+++ apps/cms-admin/src/pages/content/article/ArticleCreatePage.test.tsx
@@ -3,8 +3,7 @@ import { useParams } from 'react-router-dom'
 import { fireEvent, render, screen, waitFor } from '@testing-library/react'
 import type { MockedFunction } from 'vitest'
 import { api } from 'src/client'
-import type { ArticleEntity, HttpResponse } from 'src/client/cms/cms-api'
-import type { PaginatedEntity } from 'src/client/type'
+import type { ArticleEntity, HttpResponse, PaginatedEntity } from 'src/client/cms/cms-api'
 import { articleFixture, paginatedMetadataFixture } from 'src/fixtures'
 import ArticleCreatePage from './ArticleCreatePage'
