+++ apps/cms-admin/src/pages/content/article/ArticleEditPage.test.tsx
@@ -3,8 +3,7 @@ import { useParams } from 'react-router-dom'
 import { act, fireEvent, render, screen, waitFor } from '@testing-library/react'
 import type { MockedFunction } from 'vitest'
 import { api } from 'src/client'
-import type { ArticleEntity, HttpResponse, TagEntity } from 'src/client/cms/cms-api'
-import type { PaginatedEntity } from 'src/client/type'
+import type { ArticleEntity, HttpResponse, PaginatedEntity, TagEntity } from 'src/client/cms/cms-api'
 import { paginatedMetadataFixture, tagFixture } from 'src/fixtures'
 import ArticleEditPage from './ArticleEditPage'
